+++ Changelog.md
@@ -1,5 +1,6 @@
 # 11.0.2-20240921-2
 * Don't hardcode Accept and Cancel.
+* Persist window size across logins.
 
 # 11.0.2-20240921-1
 * Fix minimap icon disable/enable.